+++ assets/javascripts/kitten/components/images/image-cropper/image-cropper.stories.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { ImageCropper } from './index'
+import { select } from '@storybook/addon-knobs'
 
 export default {
   title: 'Images/Cropper',
@@ -9,7 +10,12 @@ export default {
 export const Default = args => {
   return (
     <div className="k-u-margin-vertical-quadruple k-u-margin-horizontal-quadruple">
-      <ImageCropper {...args} />
+      <ImageCropper
+        {...args}
+        buttonProps={{
+          variant: select('Variant', ['andromeda', 'orion'], 'andromeda'),
+        }}
+      />
     </div>
   )
 }
@@ -17,7 +23,13 @@ export const Default = args => {
 export const WithImage = () => {
   return (
     <div className="k-u-margin-vertical-quadruple k-u-margin-horizontal-quadruple">
-      <ImageCropper imageSrc="/kitten.jpg" fileName="Steven" />
+      <ImageCropper
+        imageSrc="/kitten.jpg"
+        fileName="Steven Kitten"
+        buttonProps={{
+          variant: select('Variant', ['andromeda', 'orion'], 'andromeda'),
+        }}
+      />
     </div>
   )
 }